CSS Comments + Last Update
Put everything at the top because it saved a lot of copy pasting for each command i used more than once. They are in the order of me scrolling. While alphabetical would be more useful , I find that would seem like I just copied it from elsewhere.

This is the Final Update.